writeTempToken after get
