+++ api/sts_token_test.go
@@ -1,7 +1,6 @@
 package api
 
 import (
-	"errors"
 	"io/ioutil"
 	"os"
 	"path/filepath"
@@ -45,9 +44,7 @@ func TestFetchCache(t *testing.T) {
 	defer func() { stsTokenCachePath = origstsTokenCachePath }()
 
 	err := stsToken.fetchCache()
-	if assert.Error(err) {
-		assert.Equal(errors.New("no STS cache"), err)
-	}
+	assert.EqualError(err, "no STS cache")
 
 	stsToken.createCache()
 
@@ -65,7 +62,30 @@ func TestFetchCache(t *testing.T) {
 	os.Chtimes(stsTokenCachePath, atime, mtime)
 
 	err = stsToken.fetchCache()
-	if assert.Error(err) {
-		assert.Equal(errors.New("STS cache too old, will be renew"), err)
+	assert.EqualError(err, "STS cache too old, will be renew")
+}
+
+func TestWriteTempToken(t *testing.T) {
+	assert := assert.New(t)
+	tmpAWSDir, _ := ioutil.TempDir("", "TestWriteTempToken")
+	defer os.RemoveAll(tmpAWSDir)
+
+	origawsConfDir := awsConfDir
+	awsConfDir = filepath.Join(tmpAWSDir, ".aws")
+	defer func() { awsConfDir = origawsConfDir }()
+
+	sts := &StsToken{
+		AccessKeyID:     "ASIAXXXXXXXXXXXXXXX",
+		SecretAccessKey: "SAXXX",
+		SessionToken:    "STSTOKENXXX",
 	}
+	sts.writeTempToken()
+
+	buf, err := ioutil.ReadFile(filepath.Join(awsConfDir, "awslogs_creds.conf"))
+	if err != nil {
+		t.Fatal("awslogs_creds file was not created")
+	}
+
+	assert.Contains(string(buf), "ASIAX")
+	assert.Contains(string(buf), "STSTOKENXXX")
 }
